+++ plugins/one_backup_restore/config/config.exs
@@ -2,6 +2,7 @@ use Mix.Config
 
 config :unbrella, :plugins, one_backup_restore: [
   module: OneBackupRestore,
+  application: OneBackupRestore.Application,
 ]
 
 import_config "#{Mix.env}.exs"