+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # InfinityOne Changelog
 
+## 1.0.0.beta13 (2018-04-xx)
+
+### Enhancements
+
+* [UCX-3907] Support only one logged in device. Logout other devices automatically
+
+### Bug Fixes
+
+* [UCX-3898] Can't unmute user from User details flex panel
+
+
+## 1.0.0.beta12 (2018-04-09)
+
+### Enhancements
+
+### Bug Fixes
+
+* [UCX-3896] Message replacement patterns with OneChat.refresh_users_status exception
+* [UCX-3897] Fix Error when deleting user on system with only one admin.
+* [UCX-3900] Should not be able to mute a user in DM channel
+
 ## 1.0.0.beta11 (2018-04-05)
 
 ### Enhancements